# apps/ts/packages/shared/openapi/bt-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "trading25-bt API",
    "description": "バックテスト実行のためのREST API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/optimize/run": {
      "post": {
        "tags": [
          "Optimization"
        ],
        "summary": "Run Optimization",
        "description": "パラメータ最適化を実行",
        "operationId": "run_optimization_api_optimize_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/optimize/jobs/{job_id}": {
      "get": {
        "tags": [
          "Optimization"
        ],
        "summary": "Get Optimization Status",
        "description": "最適化ジョブのステータスを取得",
        "operationId": "get_optimization_status_api_optimize_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/optimize/jobs/{job_id}/stream": {
      "get": {
        "tags": [
          "Optimization"
        ],
        "summary": "Stream Optimization Events",
        "description": "最適化ジョブの進捗をSSEでストリーミング\n\nArgs:\n    job_id: ジョブID",
        "operationId": "stream_optimization_events_api_optimize_jobs__job_id__stream_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/optimize/grid-configs/{strategy}": {
      "get": {
        "tags": [
          "Optimization"
        ],
        "summary": "Get Grid Config",
        "description": "Grid設定を取得",
        "operationId": "get_grid_config_api_optimize_grid_configs__strategy__get",
        "parameters": [
          {
            "name": "strategy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationGridConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Optimization"
        ],
        "summary": "Save Grid Config",
        "description": "Grid設定を保存",
        "operationId": "save_grid_config_api_optimize_grid_configs__strategy__put",
        "parameters": [
          {
            "name": "strategy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizationGridSaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationGridSaveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Optimization"
        ],
        "summary": "Delete Grid Config",
        "description": "Grid設定を削除",
        "operationId": "delete_grid_config_api_optimize_grid_configs__strategy__delete",
        "parameters": [
          {
            "name": "strategy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Grid Config Api Optimize Grid Configs  Strategy  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/optimize/html-files": {
      "get": {
        "tags": [
          "Optimization"
        ],
        "summary": "List Optimization Html Files",
        "description": "最適化結果HTMLファイル一覧を取得",
        "operationId": "list_optimization_html_files_api_optimize_html_files_get",
        "parameters": [
          {
            "name": "strategy",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strategy"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Optimization Html Files Api Optimize Html Files Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/optimize/html-files/{strategy}/{filename}": {
      "get": {
        "tags": [
          "Optimization"
        ],
        "summary": "Get Optimization Html File Content",
        "description": "最適化結果HTMLファイルのコンテンツを取得",
        "operationId": "get_optimization_html_file_content_api_optimize_html_files__strategy___filename__get",
        "parameters": [
          {
            "name": "strategy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Optimization Html File Content Api Optimize Html Files  Strategy   Filename  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Optimization"
        ],
        "summary": "Delete Optimization Html File",
        "description": "最適化結果HTMLファイルを削除\n\nArgs:\n    strategy: 戦略名\n    filename: ファイル名",
        "operationId": "delete_optimization_html_file_api_optimize_html_files__strategy___filename__delete",
        "parameters": [
          {
            "name": "strategy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HtmlFileDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/optimize/html-files/{strategy}/{filename}/rename": {
      "post": {
        "tags": [
          "Optimization"
        ],
        "summary": "Rename Optimization Html File",
        "description": "最適化結果HTMLファイルをリネーム\n\nArgs:\n    strategy: 戦略名\n    filename: 現在のファイル名\n    request: リネームリクエスト（新しいファイル名）",
        "operationId": "rename_optimization_html_file_api_optimize_html_files__strategy___filename__rename_post",
        "parameters": [
          {
            "name": "strategy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HtmlFileRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HtmlFileRenameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lab/optimize": {
      "post": {
        "tags": [
          "Lab"
        ],
        "summary": "Run Lab Optimize",
        "description": "Optuna最適化ジョブをサブミット",
        "operationId": "run_lab_optimize_api_lab_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabOptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HtmlFileDeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "削除成功フラグ"
          },
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name",
            "description": "戦略名"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "削除されたファイル名"
          }
        },
        "type": "object",
        "required": [
          "success",
          "strategy_name",
          "filename"
        ],
        "title": "HtmlFileDeleteResponse",
        "description": "HTMLファイル削除レスポンス"
      },
...
      "HtmlFileRenameRequest": {
        "properties": {
          "new_filename": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "New Filename",
            "description": "新しいファイル名（.html拡張子必須）"
          }
        },
        "type": "object",
        "required": [
          "new_filename"
        ],
        "title": "HtmlFileRenameRequest",
        "description": "HTMLファイルリネームリクエスト"
      },
      "HtmlFileRenameResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "リネーム成功フラグ"
          },
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name",
            "description": "戦略名"
          },
          "old_filename": {
            "type": "string",
            "title": "Old Filename",
            "description": "変更前のファイル名"
          },
          "new_filename": {
            "type": "string",
            "title": "New Filename",
            "description": "変更後のファイル名"
          }
        },
        "type": "object",
        "required": [
          "success",
          "strategy_name",
          "old_filename",
          "new_filename"
        ],
        "title": "HtmlFileRenameResponse",
        "description": "HTMLファイルリネームレスポンス"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "description": "ジョブステータス"
      },
...
      "LabOptimizeRequest": {
        "properties": {
          "strategy_name": {
            "type": "string",
            "minLength": 1,
            "title": "Strategy Name",
            "description": "ベース戦略名"
          },
          "trials": {
            "type": "integer",
            "maximum": 1000,
            "minimum": 10,
            "title": "Trials",
            "description": "試行回数",
            "default": 100
          },
          "sampler": {
            "type": "string",
            "enum": [
              "tpe",
              "random",
              "cmaes"
            ],
            "title": "Sampler",
            "description": "サンプラー",
            "default": "tpe"
          },
          "save": {
            "type": "boolean",
            "title": "Save",
            "description": "結果をYAMLに保存",
            "default": true
          },
          "scoring_weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoring Weights",
            "description": "スコアリング重み"
          }
        },
        "type": "object",
        "required": [
          "strategy_name"
        ],
        "title": "LabOptimizeRequest",
        "description": "Optuna最適化リクエスト"
      },
...
      "OptimizationGridConfig": {
        "properties": {
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name",
            "description": "戦略名"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "YAML文字列"
          },
          "param_count": {
            "type": "integer",
            "title": "Param Count",
            "description": "パラメータ数"
          },
          "combinations": {
            "type": "integer",
            "title": "Combinations",
            "description": "組み合わせ数"
          }
        },
        "type": "object",
        "required": [
          "strategy_name",
          "content",
          "param_count",
          "combinations"
        ],
        "title": "OptimizationGridConfig",
        "description": "Grid設定"
      },
...
      "OptimizationGridSaveRequest": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content",
            "description": "YAML文字列"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "OptimizationGridSaveRequest",
        "description": "Grid設定保存リクエスト"
      },
      "OptimizationGridSaveResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "保存成功フラグ"
          },
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name",
            "description": "戦略名"
          },
          "param_count": {
            "type": "integer",
            "title": "Param Count",
            "description": "パラメータ数"
          },
          "combinations": {
            "type": "integer",
            "title": "Combinations",
            "description": "組み合わせ数"
          }
        },
        "type": "object",
        "required": [
          "success",
          "strategy_name",
          "param_count",
          "combinations"
        ],
        "title": "OptimizationGridSaveResponse",
        "description": "Grid設定保存レスポンス"
      },
      "OptimizationJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "ジョブID"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress",
            "description": "進捗（0.0 - 1.0）"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "ステータスメッセージ"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "作成日時"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At",
            "description": "開始日時"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "完了日時"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "エラーメッセージ"
          },
          "best_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Score",
            "description": "最良スコア"
          },
          "total_combinations": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Combinations",
            "description": "パラメータ組み合わせ総数"
          },
          "notebook_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notebook Path",
            "description": "結果Notebookパス"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "created_at"
        ],
        "title": "OptimizationJobResponse",
        "description": "最適化ジョブレスポンス"
      },
      "OptimizationRequest": {
        "properties": {
          "strategy_name": {
            "type": "string",
            "minLength": 1,
            "title": "Strategy Name",
            "description": "戦略名"
          }
        },
        "type": "object",
        "required": [
          "strategy_name"
        ],
        "title": "OptimizationRequest",
        "description": "最適化リクエスト"
      },
...
    }
  }
}